backend/requirements.txt
Flask
gunicorn
pandas
scikit-learn[alldeps]
google-generativeai
python-dotenv
flask-cors
joblib
